+++ splunk_add_on_ucc_framework/global_config.py
@@ -131,7 +131,6 @@ class GlobalConfig:
         self._expand_entities(
             self._content.get("pages", {}).get("inputs", {}).get("services")
         )
-        self._expand_entities(self._content.get("alerts"))
 
     @staticmethod
     def _expand_entities(items: Optional[List[Dict[Any, Any]]]) -> None:
@@ -206,6 +205,10 @@ class GlobalConfig:
         return self._content.get("alerts", [])
 
     @property
+    def custom_search_commands(self) -> List[Dict[str, Any]]:
+        return self._content.get("customSearchCommand", [])
+
+    @property
     def meta(self) -> Dict[str, Any]:
         return self._content["meta"]
 
@@ -266,6 +269,9 @@ class GlobalConfig:
 
     def has_alerts(self) -> bool:
         return bool(self.alerts)
+
+    def has_custom_search_commands(self) -> bool:
+        return bool(self.custom_search_commands)
 
     def has_dashboard(self) -> bool:
         return self.dashboard is not None