+++ ui/src/components/ConfigurationFormView.jsx
@@ -11,6 +11,7 @@ import { MODE_CONFIG } from '../constants/modes';
 import { WaitSpinnerWrapper } from './table/CustomTableStyle';
 import { PAGE_CONF } from '../constants/pages';
 import PageContext from '../context/PageContext';
+import CustomComponentContext from '../context/CustomComponentContext';
 
 const ButtonWrapper = styled.div`
     margin-left: 270px !important;
@@ -72,20 +73,26 @@ function ConfigurationFormView({ serviceName }) {
     // Ref is used here to call submit method of form only
     return Object.keys(currentServiceState).length ? (
         <>
-            <PageContext.Consumer>
-                {(pageContext) => (
-                    <BaseFormView // nosemgrep: typescript.react.security.audit.react-no-refs.react-no-refs
-                        ref={form}
-                        page={PAGE_CONF}
-                        stanzaName={serviceName}
-                        serviceName="settings"
-                        mode={MODE_CONFIG}
-                        currentServiceState={currentServiceState}
-                        handleFormSubmit={handleFormSubmit}
-                        pageContext={pageContext}
-                    />
+            <CustomComponentContext.Consumer>
+                {(customComponentContext) => (
+                    <PageContext.Consumer>
+                        {(pageContext) => (
+                            <BaseFormView // nosemgrep: typescript.react.security.audit.react-no-refs.react-no-refs
+                                ref={form}
+                                page={PAGE_CONF}
+                                stanzaName={serviceName}
+                                serviceName="settings"
+                                mode={MODE_CONFIG}
+                                currentServiceState={currentServiceState}
+                                handleFormSubmit={handleFormSubmit}
+                                pageContext={pageContext}
+                                customComponentContext={customComponentContext}
+                            />
+                        )}
+                    </PageContext.Consumer>
                 )}
-            </PageContext.Consumer>
+            </CustomComponentContext.Consumer>
+
             <ButtonWrapper>
                 <UCCButton
                     className="saveBtn"