+++ ui/src/components/CustomControl/CustomControl.tsx
@@ -7,6 +7,7 @@ import { CustomValidatorFunc, UtilBaseForm } from '../../types/components/BaseFo
 import { invariant } from '../../util/invariant';
 import { CustomControlConstructor } from './CustomControlBase';
 import { ControlData } from './CustomControl.types';
+import CustomComponentContext from '../../context/CustomComponentContext';
 
 interface Props {
     data: ControlData;
@@ -22,13 +23,20 @@ interface State {
 }
 
 class CustomControl extends React.Component<Props, State> {
+    static contextType = CustomComponentContext;
+
     static loadCustomControl = (
         module: string,
         type: string,
-        appName: string
+        appName: string,
+        context?: React.ContextType<typeof CustomComponentContext>
     ): Promise<CustomControlConstructor> =>
         new Promise((resolve) => {
-            if (type === 'external') {
+            const customComp = context?.[module];
+            if (customComp?.type === 'control') {
+                const Control = customComp.component;
+                resolve(Control);
+            } else if (type === 'external') {
                 import(/* @vite-ignore */ `${getBuildDirPath()}/custom/${module}.js`).then(
                     async (external) => {
                         const Control = external.default as CustomControlConstructor;
@@ -36,13 +44,12 @@ class CustomControl extends React.Component<Props, State> {
                     }
                 );
             } else {
-                // @ts-expect-error typeof __non_webpack_require__ is not known during bundle
-                __non_webpack_require__(
-                    [`app/${appName}/js/build/custom/${module}`],
-                    (Control: CustomControlConstructor) => {
-                        resolve(Control);
-                    }
-                );
+                // eslint-disable-next-line import/no-dynamic-require, global-require
+                require([`app/${appName}/js/build/custom/${module}`], (
+                    Control: CustomControlConstructor
+                ) => {
+                    resolve(Control);
+                });
             }
         });
 
@@ -65,7 +72,8 @@ class CustomControl extends React.Component<Props, State> {
         CustomControl.loadCustomControl(
             this.props.controlOptions.src,
             this.props.controlOptions.type,
-            appName
+            appName,
+            this.context
         ).then((Control) => {
             invariant(this.el !== undefined, 'Element should be defined');
             const customControl = new Control(