+++ ui/src/components/CustomTab/CustomTab.tsx
@@ -1,15 +1,18 @@
-import React, { useEffect, useRef, useState } from 'react';
+import React, { useContext, useEffect, useRef, useState } from 'react';
 import { _ } from '@splunk/ui-utils/i18n';
 import { getUnifiedConfigs } from '../../util/util';
 import { getBuildDirPath } from '../../util/script';
 import { CustomTabConstructor } from './CustomTabBase';
 import { Tab } from './CustomTab.types';
+import CustomComponentContext from '../../context/CustomComponentContext';
 
 interface CustomTabProps {
     tab: Tab;
 }
 
 const CustomTab: React.FC<CustomTabProps> = ({ tab }) => {
+    const customCompontentContext = useContext(CustomComponentContext);
+
     const [loading, setLoading] = useState(true);
     const divRef = useRef<HTMLDivElement>(null);
 
@@ -18,7 +21,11 @@ const CustomTab: React.FC<CustomTabProps> = ({ tab }) => {
 
     const loadCustomTab = (): Promise<CustomTabConstructor> =>
         new Promise((resolve) => {
-            if (tab.customTab?.type === 'external') {
+            const customComp = customCompontentContext?.[tab?.customTab?.src];
+            if (customComp?.type === 'tab') {
+                const Control = customComp.component;
+                resolve(Control);
+            } else if (tab.customTab?.type === 'external') {
                 import(
                     /* @vite-ignore */ `${getBuildDirPath()}/custom/${tab.customTab.src}.js`
                 ).then((external) => {
@@ -26,11 +33,12 @@ const CustomTab: React.FC<CustomTabProps> = ({ tab }) => {
                     resolve(Control);
                 });
             } else {
-                // @ts-expect-error should be exported to other js module and imported here
-                __non_webpack_require__(
-                    [`app/${appName}/js/build/custom/${tab.customTab?.src}`],
-                    (Control: CustomTabConstructor) => resolve(Control)
-                );
+                // eslint-disable-next-line import/no-dynamic-require, global-require
+                require([`app/${appName}/js/build/custom/${tab.customTab!.src}`], (
+                    loadedCustomtab
+                ) => {
+                    resolve(loadedCustomtab);
+                });
             }
         });