+++ ui/src/components/table/CustomTableRow.tsx
@@ -1,4 +1,4 @@
-import React, { ReactElement, useCallback, useState } from 'react';
+import React, { ReactElement, useCallback, useContext, useState } from 'react';
 
 import WaitSpinner from '@splunk/react-ui/WaitSpinner';
 import Switch from '@splunk/react-ui/Switch';
@@ -17,6 +17,7 @@ import { getTableCellValue } from './table.utils';
 import AcceptModal from '../AcceptModal/AcceptModal';
 import { RowDataFields } from '../../context/TableContext';
 import CustomTableCell from './CustomTableCell';
+import CustomComponentContext from '../../context/CustomComponentContext';
 
 const TableCellWrapper = styled(Table.Cell)`
     padding: 2px;
@@ -63,19 +64,23 @@ function CustomTableRow(props: CustomTableRowProps) {
     } = props;
 
     const [displayAcceptToggling, setDisplayAcceptToggling] = useState(false);
+    const componentContext = useContext(CustomComponentContext);
 
     const toggleRef = React.createRef<HTMLDivElement>();
 
-    const getCustomCell = (customRow: RowDataFields, header: CellHeader) =>
-        header.customCell?.src &&
-        header.customCell?.type &&
-        React.createElement(CustomTableCell, {
-            serviceName: row.serviceName,
-            field: header.field,
-            row: customRow,
-            fileName: header.customCell.src,
-            type: header.customCell.type,
-        });
+    const getCustomCell = (customRow: RowDataFields, header: CellHeader) => {
+        return (
+            header.customCell?.src &&
+            React.createElement(CustomTableCell, {
+                serviceName: row.serviceName,
+                field: header.field,
+                row: customRow,
+                fileName: header.customCell.src,
+                type: header.customCell.type,
+                customComponentContext: componentContext,
+            })
+        );
+    };
 
     const rowActionsPrimaryButton = useCallback(
         (selectedRow: RowDataFields, header: CellHeader) => (