+++ docs/CHANGELOG.md
@@ -1,5 +1,36 @@
 # Changelog
 
+# [5.66.0](https://github.com/splunk/addonfactory-ucc-generator/compare/v5.65.0...v5.66.0) (2025-06-03)
+
+
+### Bug Fixes
+
+* custom tab without entity error ([#1718](https://github.com/splunk/addonfactory-ucc-generator/issues/1718)) ([9c38fd4](https://github.com/splunk/addonfactory-ucc-generator/commit/9c38fd4a2f12e8bd35e7d87f5a5268938e78603b))
+* handling of non external custom elements ([#1766](https://github.com/splunk/addonfactory-ucc-generator/issues/1766)) ([2342a05](https://github.com/splunk/addonfactory-ucc-generator/commit/2342a050aeecca9a34f15ca3e0c0c30ad929f771))
+* remove support for interval and index type of entities for alerts ([#1757](https://github.com/splunk/addonfactory-ucc-generator/issues/1757)) ([49c04be](https://github.com/splunk/addonfactory-ucc-generator/commit/49c04be804bcd27f820d2bb8d77b0674499b513c))
+* set host and port for proxy as required ([#1755](https://github.com/splunk/addonfactory-ucc-generator/issues/1755)) ([4bafe79](https://github.com/splunk/addonfactory-ucc-generator/commit/4bafe79e8ffeebe0b10ed294157ec06b387cf0ce))
+* **ui lib:** do not copy files from lib while building ([#1756](https://github.com/splunk/addonfactory-ucc-generator/issues/1756)) ([3b433ee](https://github.com/splunk/addonfactory-ucc-generator/commit/3b433eebbb2c677c70009581389a3537423bc8a4))
+
+
+### Features
+
+* **playground:** api calls as success to enable internal data flow ([#1731](https://github.com/splunk/addonfactory-ucc-generator/issues/1731)) ([6170551](https://github.com/splunk/addonfactory-ucc-generator/commit/6170551f539c150b4a88e63b57d4ab3e7b2569ad))
+
+# [5.65.0](https://github.com/splunk/addonfactory-ucc-generator/compare/v5.64.0...v5.65.0) (2025-05-15)
+
+
+### Bug Fixes
+
+* **Oauth Client Credentials:** OpenAPI fix ([#1748](https://github.com/splunk/addonfactory-ucc-generator/issues/1748)) ([fe2eb86](https://github.com/splunk/addonfactory-ucc-generator/commit/fe2eb86c37985bc78890871f0ebe6867622b6fac))
+
+
+### Features
+
+* add ucc-gen-ui cli command ([#1705](https://github.com/splunk/addonfactory-ucc-generator/issues/1705)) ([d9c3210](https://github.com/splunk/addonfactory-ucc-generator/commit/d9c3210eee083cd527c33e6595bd98e259d53e92))
+* base code for supporting custom search command  ([#1693](https://github.com/splunk/addonfactory-ucc-generator/issues/1693)) ([03cb0d2](https://github.com/splunk/addonfactory-ucc-generator/commit/03cb0d2602cd4a63e98a7fcb3c1525d501d4d26c)), closes [#1694](https://github.com/splunk/addonfactory-ucc-generator/issues/1694) [#1695](https://github.com/splunk/addonfactory-ucc-generator/issues/1695) [#1697](https://github.com/splunk/addonfactory-ucc-generator/issues/1697)
+* custom react component via context ([#1706](https://github.com/splunk/addonfactory-ucc-generator/issues/1706)) ([1a65cf0](https://github.com/splunk/addonfactory-ucc-generator/commit/1a65cf0f0a06e2633660cc83f6aa4ef04dffa8fd))
+* update base.html if it is incorrect ([#1750](https://github.com/splunk/addonfactory-ucc-generator/issues/1750)) ([05066ca](https://github.com/splunk/addonfactory-ucc-generator/commit/05066ca8ba5284c8ec120c8731817d67a2edf9c0))
+
 # [5.64.0](https://github.com/splunk/addonfactory-ucc-generator/compare/v5.63.0...v5.64.0) (2025-05-08)